+++ apis/vara-eth/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Changed
 
 - Use raw data in `signMessage` method in https://github.com/gear-tech/gear-js/pull/2254
+- Use blake2 for both messageId and tx hash in https://github.com/gear-tech/gear-js/pull/2254
 
 ### Added